+++ src/openmap/com/bbn/openmap/graphicLoader/GLPoint.java
@@ -14,8 +14,8 @@
 // 
 // $Source: /cvs/distapps/openmap/src/openmap/com/bbn/openmap/graphicLoader/GLPoint.java,v $
 // $RCSfile: GLPoint.java,v $
-// $Revision: 1.4 $
-// $Date: 2005/12/09 21:09:06 $
+// $Revision: 1.5 $
+// $Date: 2009/01/21 01:24:42 $
 // $Author: dietrick $
 // 
 // **********************************************************************
@@ -62,7 +62,7 @@ public class GLPoint extends OMPoint implements ActionListener, FocusListener {
     protected String name;
     protected int height = 0;// meters.
 
-    public GLPoint(float lat, float lon, int radius, boolean isOval) {
+    public GLPoint(double lat, double lon, int radius, boolean isOval) {
         super(lat, lon, radius);
         setOval(isOval);
     }
